requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
lxml==4.8.0
python-docx==0.8.11
soupsieve==2.3.1
